Setting up browser navigation and improving event edit page
