+++ app/src/main/java/com/example/scanpal/Fragments/BrowseYourEventFragment.java
@@ -6,9 +6,8 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.GridView;
-import android.widget.Spinner;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
@@ -18,13 +17,14 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 
-import com.example.scanpal.Models.Event;
-import com.example.scanpal.Controllers.EventController;
 import com.example.scanpal.Adapters.EventGridAdapter;
+import com.example.scanpal.Callbacks.EventFetchByUserCallback;
 import com.example.scanpal.Callbacks.EventsFetchCallback;
-import com.example.scanpal.R;
-import com.example.scanpal.Controllers.UserController;
 import com.example.scanpal.Callbacks.UserSignedUpCallback;
+import com.example.scanpal.Controllers.EventController;
+import com.example.scanpal.Controllers.UserController;
+import com.example.scanpal.Models.Event;
+import com.example.scanpal.R;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.firebase.firestore.FirebaseFirestore;
 
@@ -36,7 +36,7 @@ import java.util.concurrent.CountDownLatch;
  * Fragment for displaying a list of events. Allows users to navigate to event details,
  * add new events, and scan QR codes for event-related actions.
  */
-public class EventPageFragment extends Fragment {
+public class BrowseYourEventFragment extends Fragment {
 
     //necessary to request user for notification perms
     private final ActivityResultLauncher<String> requestPermissionLauncher =
@@ -47,44 +47,38 @@ public class EventPageFragment extends Fragment {
             });
     protected List<Event> eventsList = new ArrayList<>();
     protected List<Event> allEvents = new ArrayList<>();
+
+    private GridView gridView;
     private EventGridAdapter adapter;
     private EventController eventController;
+
 
     /**
      * Default constructor. Initializes the fragment.
      */
-    public EventPageFragment() {
+    public BrowseYourEventFragment() {
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.events_page, container, false);
+        View view = inflater.inflate(R.layout.browse_events, container, false);
+
+
+        ((TextView) view.findViewById(R.id.event_page_title)).setText("Your Events");
 
         adapter = new EventGridAdapter(getContext(), new ArrayList<>());
-        GridView gridView = view.findViewById(R.id.event_grid);
+        gridView = view.findViewById(R.id.event_grid);
         gridView.setAdapter(adapter);
 
         // init eventController
         eventController = new EventController();
 
-        Spinner spinner = view.findViewById(R.id.browser_select);
-        // Create an ArrayAdapter using the string array and a default spinner layout.
-        ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(
-                requireContext(),
-                R.array.browser_select_array,
-                android.R.layout.simple_spinner_item
-        );
-        // Specify the layout to use when the list of choices appears.
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-        // Apply the adapter to the spinner.
-        spinner.setAdapter(adapter);
-
-        fetchAllEvents();
+        fetchYourEvents();
 
         // Set up button to add new events.
         FloatingActionButton addEventButton = view.findViewById(R.id.button_add_event);
         addEventButton.setOnClickListener(v -> {
-            NavController navController = NavHostFragment.findNavController(EventPageFragment.this);
+            NavController navController = NavHostFragment.findNavController(BrowseYourEventFragment.this);
             navController.navigate(R.id.addEvent);
         });
 
@@ -92,7 +86,7 @@ public class EventPageFragment extends Fragment {
             Event event = allEvents.get(position);
             Bundle bundle = new Bundle();
             bundle.putString("event_id", event.getId());
-            NavHostFragment.findNavController(this).navigate(R.id.select_event, bundle);
+            NavHostFragment.findNavController(this).navigate(R.id.eventDetailsPage, bundle);
         });
 
         askNotificationPermission();
@@ -102,56 +96,62 @@ public class EventPageFragment extends Fragment {
     }
 
     /**
-     * Fetches all events from the data source and updates the grid view accordingly.
+     * Fetches all events organized by the user from the data source and updates the grid view accordingly.
      * Applies filters based on user sign-up status and sorts events.
      */
-    private void fetchAllEvents() {
-        UserController userController = new UserController(FirebaseFirestore.getInstance(), this.getContext());
-        eventController.fetchAllEvents(new EventsFetchCallback() {
-            @Override
-            public void onSuccess(List<Event> events) {
-                allEvents.clear();
-                allEvents.addAll(events);
+    private void fetchYourEvents() {
+        UserController userController = new UserController(FirebaseFirestore.getInstance(), getContext());
+        String username = userController.fetchStoredUsername();
+        if (username != null) {
+            // Use EventController to fetch user-specific events
+            eventController.getEventsByUser(new View(getContext()), new EventFetchByUserCallback() {
+                @Override
+                public void onSuccess(List<Event> events) {
+                    allEvents.clear();
+                    allEvents.addAll(events);
 
-                CountDownLatch latch = new CountDownLatch(allEvents.size());
+                    CountDownLatch latch = new CountDownLatch(allEvents.size());
 
-                for (Event event : allEvents) {
-                    userController.isUserSignedUp(userController.fetchStoredUsername(), event.getId(), new UserSignedUpCallback() {
-                        @Override
-                        public void onResult(boolean isSignedUp) {
-                            event.setUserSignedUp(isSignedUp);
-                            latch.countDown();
+                    for (Event event : allEvents) {
+                        userController.isUserSignedUp(username, event.getId(), new UserSignedUpCallback() {
+                            @Override
+                            public void onResult(boolean isSignedUp) {
+                                event.setUserSignedUp(isSignedUp);
+                                latch.countDown();
+                            }
+
+                            @Override
+                            public void onError(Exception e) {
+                                latch.countDown();
+                            }
+                        });
+                    }
+
+                    new Thread(() -> {
+                        try {
+                            latch.await();
+                            requireActivity().runOnUiThread(() -> {
+                                allEvents.sort((o1, o2) -> Boolean.compare(o2.isUserSignedUp(), o1.isUserSignedUp()));
+                                adapter.setEvents(allEvents);
+                            });
+                        } catch (InterruptedException ignored) {
                         }
-
-                        @Override
-                        public void onError(Exception e) {
-                            latch.countDown();
-                        }
-                    });
+                    }).start();
                 }
 
-                new Thread(() -> {
-                    try {
-                        latch.await();
-                        requireActivity().runOnUiThread(() -> {
-                            allEvents.sort((o1, o2) -> Boolean.compare(o2.isUserSignedUp(), o1.isUserSignedUp()));
-                            adapter.setEvents(allEvents);
-                        });
-                    } catch (InterruptedException ignored) {
-                    }
-                }).start();
-            }
-
-            @Override
-            public void onError(Exception e) {
-                Toast.makeText(getContext(), "Error fetching all events.", Toast.LENGTH_SHORT).show();
-            }
-        });
+                @Override
+                public void onError(Exception e) {
+                    Toast.makeText(getContext(), "Error fetching all events.", Toast.LENGTH_SHORT).show();
+                }
+            });
+        } else {
+            Toast.makeText(getContext(), "Username not found", Toast.LENGTH_SHORT).show();
+        }
     }
 
     /**
      * Fetches events from the data source and updates the grid view.
-     * This method differs from fetchAllEvents by focusing on updating the existing list.
+     * This method differs from fetchYourEvents by focusing on updating the existing list.
      */
     private void fetchEventsAndUpdateGrid() {
         EventController eventController = new EventController();
