+++ app/src/main/res/values/strings.xml
@@ -6,9 +6,4 @@
     <string name="continueButton">Continue</string>
     <string name="addUserImage">User Image</string>
     <string name="title_activity_maps">MapsActivity</string>
-    <string-array name="browser_select_array">
-        <item>Events Browser</item>
-        <item>Image Browser</item>
-        <item>Profile Browser</item>
-    </string-array>
 </resources>
